feat(T-10): 工程门面收尾 — README/SnakeServer/W4/归档/Board 测试补链
- README.md: 中英双语，模块表/构建/路线图/许可证
- SnakeServer.exe: 空壳 main_server.cpp（LOG_INFO 后 exit 0）
- 编译警告: /W4 (MSVC) / -Wall -Wextra -Wpedantic (GCC)，3 处修复 + gtest /W0 抑制
- 文档归档: plan.md → archive/plan-v1.md，design.md 单源声明
- test_board.cpp 补链 (AP-5/TEST-3 关闭)，46/46 全绿 (21.3s)

Phase-2 全部完成。

任务卡: docs/tasks/T-10-engineering-polish.md
审查通过: 2026-05-19

+++ src/ui/GameScene.cpp
@@ -41,7 +41,7 @@ void GameScene::syncFromState(const GameState& state) {
             removeItem(item);
             delete item;
         }
-        for (size_t j = 0; j < body.size(); ++j) {
+        for (int j = 0; j < static_cast<int>(body.size()); ++j) {
             items[j]->setPos(cellToPixel(body[j].x), cellToPixel(body[j].y));
             items[j]->setIsHead(j == 0);
         }
